section3

+++ ch1/section2/src/Rational.java
@@ -10,7 +10,7 @@ public class Rational {
     }
 
     public Rational(int numerator, int denominator) {
-        this((long) numerator, (long) denominator);
+        this(numerator, (long) denominator);
     }
